1. Got the maintainer of google-trends-api to fix the broken url 2. Using socket-io just because 3. Refactored scoring codes out
