+++ app.js
@@ -4,18 +4,20 @@ const express = require("express");
 const http = require("http");
 const path = require('path');
 const bodyParser = require("body-parser");
+const session = require('express-session');
 const firebase = require('firebase');
 const webpack = require('webpack');
 const webpackConfig = require('./webpack.config');
 const httpProxy = require('http-proxy');
+const app = express();
+const server = http.createServer(app);
+const io = require('socket.io')(server);
+
 const proxy = httpProxy.createProxyServer({
   ws: true
 })
 
 const { app: appConf, firebase: firebaseConf } = require("./config");
-
-const app = express();
-const server = http.createServer(app);
 
 const isDeveloping = process.env.NODE_ENV !== 'production';
 const port = isDeveloping ? appConf.port.dev : appConf.port.deploy;
@@ -51,13 +53,20 @@ firebase.initializeApp(firebaseConf);
 // const db = firebase.database();
 // const patternRef = db.ref("/patterns");
 
+require('babel-core/register');
+
 server.listen(port);
 
-require('babel-core/register');
 app.set('views', path.join(__dirname, 'template'));
 app.set('view engine', 'jade');
 app.use(express.static(path.join(__dirname, 'public')));
-
+app.use(session({
+  secret: "pleasedothis",
+  name: 'this_cookie',
+  proxy: true,
+  resave: true,
+  saveUninitialized: true
+}));
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({
   extended: true
@@ -66,6 +75,7 @@ app.use(bodyParser.urlencoded({
 const scorer = require('./script/score');
 
 require("./script/routes")(app, scorer);
+require("./script/socket")(io, scorer);
 
 proxy.on('error', (e) => {
   console.log('Could not connect to proxy, please try again...')
