+++ package.json
@@ -14,18 +14,27 @@
   },
   "dependencies": {
     "alexa-traffic-rank": "^1.0.1",
-    "babel": "^6.3.13",
-    "babel-core": "^6.3.21",
-    "babel-loader": "^6.2.0",
-    "babel-preset-es2015": "^6.3.13",
-    "babel-preset-react": "^6.3.13",
+    "babel-core": "^6.4.5",
+    "babel-loader": "6.2.9",
+    "babel-plugin-react": "^1.0.0",
+    "babel-plugin-react-transform": "^2.0.2",
+    "babel-plugin-transform-class-properties": "^6.22.0",
+    "babel-plugin-transform-decorators-legacy": "^1.3.4",
+    "babel-plugin-transform-runtime": "^6.22.0",
+    "babel-polyfill": "^6.23.0",
+    "babel-preset-es2015": "^6.22.0",
+    "babel-preset-react": "^6.22.0",
+    "babel-preset-react-hmre": "^1.1.1",
+    "babel-preset-stage-0": "^6.22.0",
+    "babel-runtime": "^6.22.0",
     "body-parser": "^1.16.0",
     "cheerio": "^0.22.0",
     "css-loader": "^0.23.0",
     "denodeify": "^1.2.1",
     "express": "^4.13.1",
+    "express-session": "^1.15.1",
     "firebase": "^3.6.9",
-    "google-trends-api": "^4.0.1",
+    "google-trends-api": "^4.0.3",
     "jade": "~1.11.0",
     "jsx-loader": "^0.13.2",
     "node-jsx": "^0.13.3",
@@ -34,8 +43,15 @@
     "react": "^0.14.3",
     "react-dom": "^0.14.3",
     "reactify": "^1.1.1",
+    "socket.io": "^1.7.3",
+    "socket.io-client": "^1.7.3",
     "style-loader": "^0.13.0",
     "webpack": "^1.12.9",
-    "webpack-dev-server": "^1.9.0"
+    "webpack-dev-server": "1.16.2"
+  },
+  "devDependencies": {
+    "babel-plugin-transform-class-properties": "^6.23.0",
+    "babel-plugin-transform-es2015-arrow-functions": "^6.22.0",
+    "babel-preset-react": "^6.23.0"
   }
 }
