+++ static/jsx/homePageAnchor.jsx
@@ -5,7 +5,7 @@ import React from 'react';
 import ReactDom from 'react-dom';
 import HomePage from './pages/homePage.jsx';
 
-let mountNode = document.getElementById("mount-point");
+const mountNode = document.getElementById("mount-point");
 
 ReactDom.render(
   <HomePage app_props = { window.APP_PROPS }/>,
